Move tests from visitor to iterator
