+++ ethexe/db/src/visitor.rs
@@ -288,392 +288,3 @@ pub fn walk(visitor: &mut impl DatabaseVisitor, node: impl Into<Node>) {
         .start(node.into())
         .for_each(|node| visit_node(visitor, node));
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{Database, iterator::DatabaseIteratorError};
-    use ethexe_common::StateHashWithQueueSize;
-    use ethexe_runtime_common::state::HashOf;
-    use gprimitives::MessageId;
-    use std::collections::BTreeMap;
-
-    #[derive(Debug)]
-    struct TestVisitor {
-        db: Database,
-        visited_code_ids: Vec<CodeId>,
-        visited_program_ids: Vec<ActorId>,
-        visited_payloads: Vec<Payload>,
-        errors: Vec<DatabaseIteratorError>,
-    }
-
-    impl TestVisitor {
-        fn new() -> Self {
-            Self {
-                db: Database::memory(),
-                visited_code_ids: vec![],
-                visited_program_ids: vec![],
-                visited_payloads: vec![],
-                errors: vec![],
-            }
-        }
-    }
-
-    impl DatabaseVisitor for TestVisitor {
-        fn db(&self) -> &dyn DatabaseIteratorStorage {
-            &self.db
-        }
-
-        fn clone_boxed_db(&self) -> Box<dyn DatabaseIteratorStorage> {
-            Box::new(self.db.clone())
-        }
-
-        fn on_db_error(&mut self, error: DatabaseIteratorError) {
-            self.errors.push(error);
-        }
-
-        fn visit_code_id(&mut self, code_id: CodeId) {
-            self.visited_code_ids.push(code_id);
-        }
-
-        fn visit_program_id(&mut self, program_id: ActorId) {
-            self.visited_program_ids.push(program_id);
-        }
-
-        fn visit_payload(&mut self, payload: &Payload) {
-            self.visited_payloads.push(payload.clone());
-        }
-    }
-
-    #[test]
-    fn walk_chain_basic() {
-        let mut visitor = TestVisitor::new();
-        let head = H256::from_low_u64_be(1);
-        let bottom = H256::from_low_u64_be(2);
-
-        // This will fail because we don't have the block header in the database
-        walk(&mut visitor, ChainNode { head, bottom });
-
-        // Should have attempted to visit the head block
-        assert!(!visitor.errors.is_empty());
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoBlockHeader(head))
-        );
-    }
-
-    #[test]
-    fn walk_block_with_missing_data() {
-        let mut visitor = TestVisitor::new();
-        let block = H256::from_low_u64_be(42);
-
-        walk(&mut visitor, BlockNode { block });
-
-        // Should have errors for all missing block data
-        let expected_errors = [
-            DatabaseIteratorError::NoBlockHeader(block),
-            DatabaseIteratorError::NoBlockEvents(block),
-            DatabaseIteratorError::NoBlockCommitmentQueue(block),
-            DatabaseIteratorError::NoBlockCodesQueue(block),
-            DatabaseIteratorError::NoPreviousNonEmptyBlock(block),
-            DatabaseIteratorError::NoLastCommittedBatch(block),
-            DatabaseIteratorError::NoBlockProgramStates(block),
-            DatabaseIteratorError::NoBlockSchedule(block),
-            DatabaseIteratorError::NoBlockOutcome(block),
-        ];
-
-        for expected_error in expected_errors {
-            assert!(visitor.errors.contains(&expected_error));
-        }
-    }
-
-    #[test]
-    fn test_walk_block_codes_queue() {
-        let mut visitor = TestVisitor::new();
-
-        let block = H256::random();
-        let code_id1 = CodeId::from([1u8; 32]);
-        let code_id2 = CodeId::from([2u8; 32]);
-        let mut queue = VecDeque::new();
-        queue.push_back(code_id1);
-        queue.push_back(code_id2);
-
-        walk(
-            &mut visitor,
-            BlockCodesQueueNode {
-                block,
-                block_codes_queue: queue,
-            },
-        );
-
-        assert_eq!(visitor.visited_code_ids.len(), 2);
-        assert!(visitor.visited_code_ids.contains(&code_id1));
-        assert!(visitor.visited_code_ids.contains(&code_id2));
-    }
-
-    #[test]
-    fn test_walk_block_program_states() {
-        let mut visitor = TestVisitor::new();
-
-        let block = H256::random();
-        let program_id = ActorId::from([3u8; 32]);
-        let state_hash = H256::random();
-
-        let mut block_program_states = BTreeMap::new();
-        block_program_states.insert(
-            program_id,
-            StateHashWithQueueSize {
-                hash: state_hash,
-                cached_queue_size: 0,
-            },
-        );
-
-        walk(
-            &mut visitor,
-            BlockProgramStatesNode {
-                block,
-                block_program_states,
-            },
-        );
-
-        // Should have error because program state is not in database
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoProgramState(state_hash))
-        );
-    }
-
-    #[test]
-    fn walk_program_id_missing_code() {
-        let mut visitor = TestVisitor::new();
-        let program_id = ActorId::from([5u8; 32]);
-
-        walk(&mut visitor, ProgramIdNode { program_id });
-
-        // Should have error because program code ID is not in database
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoProgramCodeId(program_id))
-        );
-    }
-
-    #[test]
-    fn test_no_code_valid_error() {
-        let code_id = CodeId::from(1);
-
-        let mut visitor = TestVisitor::new();
-        walk(&mut visitor, CodeIdNode { code_id });
-
-        let expected_errors = [
-            DatabaseIteratorError::NoCodeValid(code_id),
-            DatabaseIteratorError::NoOriginalCode(code_id),
-            DatabaseIteratorError::NoInstrumentedCode(code_id),
-            DatabaseIteratorError::NoCodeMetadata(code_id),
-        ];
-
-        for expected_error in expected_errors {
-            assert!(visitor.errors.contains(&expected_error));
-        }
-    }
-
-    #[test]
-    fn walk_scheduled_task_pause_program() {
-        let mut visitor = TestVisitor::new();
-        let program_id = ActorId::from([6u8; 32]);
-        let task = ScheduledTask::PauseProgram(program_id);
-
-        walk(&mut visitor, ScheduledTaskNode { task });
-
-        assert!(visitor.visited_program_ids.contains(&program_id));
-    }
-
-    #[test]
-    fn walk_scheduled_task_remove_code() {
-        let mut visitor = TestVisitor::new();
-        let code_id = CodeId::from([7u8; 32]);
-        let task = ScheduledTask::RemoveCode(code_id);
-
-        walk(&mut visitor, ScheduledTaskNode { task });
-
-        assert!(visitor.visited_code_ids.contains(&code_id));
-    }
-
-    #[test]
-    fn walk_scheduled_task_wake_message() {
-        let mut visitor = TestVisitor::new();
-        let program_id = ActorId::from([8u8; 32]);
-        let msg_id = MessageId::from([9u8; 32]);
-        let task = ScheduledTask::WakeMessage(program_id, msg_id);
-
-        walk(&mut visitor, ScheduledTaskNode { task });
-
-        assert!(visitor.visited_program_ids.contains(&program_id));
-    }
-
-    #[test]
-    fn test_walk_block_schedule_tasks() {
-        let mut visitor = TestVisitor::new();
-
-        let block = H256::random();
-        let program_id1 = ActorId::from([10u8; 32]);
-        let program_id2 = ActorId::from([11u8; 32]);
-        let code_id = CodeId::from([12u8; 32]);
-
-        let mut tasks = BTreeSet::new();
-        tasks.insert(ScheduledTask::PauseProgram(program_id1));
-        tasks.insert(ScheduledTask::RemoveCode(code_id));
-        tasks.insert(ScheduledTask::WakeMessage(program_id2, MessageId::zero()));
-
-        walk(
-            &mut visitor,
-            BlockScheduleTasksNode {
-                block,
-                height: 123,
-                tasks,
-            },
-        );
-
-        assert!(visitor.visited_program_ids.contains(&program_id1));
-        assert!(visitor.visited_program_ids.contains(&program_id2));
-        assert!(visitor.visited_code_ids.contains(&code_id));
-    }
-
-    #[test]
-    fn test_walk_block_schedule() {
-        let mut visitor = TestVisitor::new();
-        let block = H256::from([13u8; 32]);
-        let program_id = ActorId::from([14u8; 32]);
-
-        let mut block_schedule = BTreeMap::new();
-        let mut tasks = BTreeSet::new();
-        tasks.insert(ScheduledTask::PauseProgram(program_id));
-        block_schedule.insert(1000u32, tasks);
-
-        walk(
-            &mut visitor,
-            BlockScheduleNode {
-                block,
-                block_schedule,
-            },
-        );
-
-        assert!(visitor.visited_program_ids.contains(&program_id));
-    }
-
-    #[test]
-    fn walk_block_outcome() {
-        let mut visitor = TestVisitor::new();
-        let block = H256::from([16u8; 32]);
-        let actor_id = ActorId::from([15u8; 32]);
-        let new_state_hash = H256::random();
-
-        walk(
-            &mut visitor,
-            BlockOutcomeNode {
-                block,
-                block_outcome: BlockOutcome::Transitions(vec![StateTransition {
-                    actor_id,
-                    new_state_hash,
-                    exited: false,
-                    inheritor: Default::default(),
-                    value_to_receive: 0,
-                    value_claims: vec![],
-                    messages: vec![],
-                }]),
-            },
-        );
-
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoProgramCodeId(actor_id))
-        );
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoProgramState(new_state_hash))
-        );
-    }
-
-    #[test]
-    fn test_walk_state_transition() {
-        let mut visitor = TestVisitor::new();
-        let actor_id = ActorId::from([17u8; 32]);
-        let new_state_hash = H256::from([18u8; 32]);
-
-        let state_transition = StateTransition {
-            actor_id,
-            new_state_hash,
-            exited: false,
-            inheritor: ActorId::zero(),
-            value_to_receive: 0,
-            value_claims: Vec::new(),
-            messages: Vec::new(),
-        };
-
-        walk(&mut visitor, StateTransitionNode { state_transition });
-
-        assert!(visitor.visited_program_ids.contains(&actor_id));
-        // Should have error for missing program state
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoProgramState(new_state_hash))
-        );
-    }
-
-    #[test]
-    fn walk_state_transition_zero_state_hash() {
-        let mut visitor = TestVisitor::new();
-        let actor_id = ActorId::from([19u8; 32]);
-
-        let state_transition = StateTransition {
-            actor_id,
-            new_state_hash: H256::zero(),
-            exited: false,
-            inheritor: ActorId::zero(),
-            value_to_receive: 0,
-            value_claims: Vec::new(),
-            messages: Vec::new(),
-        };
-
-        walk(&mut visitor, StateTransitionNode { state_transition });
-
-        assert!(visitor.visited_program_ids.contains(&actor_id));
-        // Should not try to get program state for zero hash
-        assert!(!visitor.errors.iter().any(
-            |e| matches!(e, DatabaseIteratorError::NoProgramState(hash) if *hash == H256::zero())
-        ));
-    }
-
-    #[test]
-    fn walk_payload_lookup_direct() {
-        let mut visitor = TestVisitor::new();
-        let payload_data = vec![1, 2, 3, 4];
-        let payload = Payload::try_from(payload_data.clone()).unwrap();
-        let payload_lookup = PayloadLookup::Direct(payload.clone());
-
-        walk(&mut visitor, PayloadLookupNode { payload_lookup });
-
-        assert!(visitor.visited_payloads.contains(&payload));
-    }
-
-    #[test]
-    fn walk_payload_lookup_stored() {
-        let mut visitor = TestVisitor::new();
-        let payload_hash = unsafe { HashOf::<Payload>::new(H256::zero()) };
-        let payload_lookup = PayloadLookup::Stored(payload_hash);
-
-        walk(&mut visitor, PayloadLookupNode { payload_lookup });
-
-        // Should have error for missing stored payload
-        assert!(
-            visitor
-                .errors
-                .contains(&DatabaseIteratorError::NoPayload(payload_hash))
-        );
-    }
-}
